change "injected" to "web3context"
It seems like the correct definition of web3context IS injected. Previously it says injected here, but it's unclear to readers what web3context is. I think it was intended that injected was meant to be named `web3context`.  If not, then there should be a clarification as to what `web3context` actually should be in the lines that follow.